InputHandler renamed to CommandHandler (#8)
Add command was decomposed to InputHandler
