+++ src/main/java/Main.java
@@ -1,4 +1,4 @@
-import classes.console.InputHandler;
+import classes.console.CommandHandler;
 import classes.console.TextColor;
 import exceptions.NoSuchCommandException;
 import exceptions.SystemException;
@@ -10,15 +10,18 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        InputHandler inputHandler = new InputHandler();
+        CommandHandler commandHandler = new CommandHandler();
         while (true) {
             System.out.print(TextColor.green("> "));
             try {
                 String inputString = scanner.nextLine();
                 String commandName = inputString.split(" ")[0];
-                String commandArgument = inputString.split(" ")[1];
+                String commandArgument = null;
+                if (inputString.split(" ").length > 1) {
+                    commandArgument = inputString.split(" ")[1];
+                }
                 if (!commandName.isBlank()) {
-                    Commandable command = inputHandler.getCommand(commandName);
+                    Commandable command = commandHandler.getCommand(commandName);
                     command.execute(commandArgument);
                 }
             } catch (NoSuchCommandException e) {
